ticket #840: bad deleted directory with jelix scripts. p=me+julieni  trunk+1.1.x
--HG--
extra : convert_revision : svn%3Ab2e90908-df08-0410-9efa-d317cf624d33/trunk%401318

+++ lib/jelix-scripts/commands/cleartemp.cmd.php
@@ -27,19 +27,52 @@ class cleartempCommand extends JelixScriptCommand {
 
     public function run(){
         try {
+            if (!defined('JELIX_APP_TEMP_PATH')) {
+                echo "Error: JELIX_APP_TEMP_PATH is not defined\n";
+                exit(1);
+            }
+            if (JELIX_APP_TEMP_PATH == DIRECTORY_SEPARATOR || JELIX_APP_TEMP_PATH == '' || JELIX_APP_TEMP_PATH == '/') {
+                echo "Error: bad path in JELIX_APP_TEMP_PATH, it is equals to '".JELIX_APP_TEMP_PATH."' !!\n";
+                echo "       Jelix cannot clear the content of the temp directory.\n";
+                echo "       Correct the path in JELIX_APP_TEMP_PATH or create the directory you\n";
+                echo "       indicated into JELIX_APP_TEMP_PATH.\n";
+                exit(1);
+            }
             jFile::removeDir(JELIX_APP_TEMP_PATH, false);
-	    jFile::removeDir(JELIX_APP_REAL_TEMP_PATH, false);
-	    if (defined('JELIX_APP_TEMP_CLI_PATH'))
-		jFile::removeDir(JELIX_APP_TEMP_CLI_PATH, false);
+
+
+            if (!defined('JELIX_APP_REAL_TEMP_PATH')) {
+                echo "Error: JELIX_APP_REAL_TEMP_PATH is not defined\n";
+                exit(1);
+            }
+            if (JELIX_APP_REAL_TEMP_PATH == DIRECTORY_SEPARATOR || JELIX_APP_REAL_TEMP_PATH == '' || JELIX_APP_REAL_TEMP_PATH == '/') {
+                echo "Error: bad path in JELIX_APP_REAL_TEMP_PATH, it is equals to '".JELIX_APP_REAL_TEMP_PATH."' !!\n";
+                echo "       Jelix cannot clear the content of the temp directory.\n";
+                echo "       Correct the path in JELIX_APP_REAL_TEMP_PATH or create the directory you\n";
+                echo "       indicated into JELIX_APP_REAL_TEMP_PATH.\n";
+                exit(1);
+            }
+            jFile::removeDir(JELIX_APP_REAL_TEMP_PATH, false);
+
+
+            if (defined('JELIX_APP_TEMP_CLI_PATH')){
+                if (JELIX_APP_TEMP_CLI_PATH == DIRECTORY_SEPARATOR || JELIX_APP_TEMP_CLI_PATH == '' || JELIX_APP_TEMP_CLI_PATH == '/') {
+                    echo "Error: bad path in JELIX_APP_TEMP_CLI_PATH, it is equals to '".JELIX_APP_TEMP_CLI_PATH."' !!\n";
+                    echo "       Jelix cannot clear the content of the temp directory.\n";
+                    echo "       Correct the path in JELIX_APP_TEMP_CLI_PATH or create the directory you\n";
+                    echo "       indicated into JELIX_APP_TEMP_CLI_PATH.\n";
+                    exit(1);
+                }
+                jFile::removeDir(JELIX_APP_TEMP_CLI_PATH, false);
+            }
         }
         catch (Exception $e) {
             if(MESSAGE_LANG == 'fr')
-        	   echo "Un ou plusieurs répertoires n'ont pas pu être supprimés.\n" .
+               echo "Un ou plusieurs répertoires n'ont pas pu être supprimés.\n" .
                     "Message d'erreur : " . $e->getMessage()."\n";
             else
-        	   echo "One or more directories couldn't be deleted.\n" .
+               echo "One or more directories couldn't be deleted.\n" .
                     "Error: " . $e->getMessage()."\n";
         }
     }
 }
-
